Eliminate redundant (and slow) tests
