+++ modules/importer/src/runner/cve/handler.rs
@@ -92,43 +92,3 @@ where
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod test {
-    use crate::{
-        model::DEFAULT_SOURCE_CVEPROJECT,
-        runner::common::walker::{git_reset, Continuation, GitWalker},
-    };
-    use std::path::PathBuf;
-
-    /// test CVE walker, runs for a long time
-    #[test_log::test(tokio::test)]
-    async fn test_cve_walker() -> Result<(), anyhow::Error> {
-        let path = PathBuf::from(format!(
-            "{}target/test.data/test_cve_walker.git",
-            env!("CARGO_WORKSPACE_ROOT")
-        ));
-        if path.exists() {
-            std::fs::remove_dir_all(path.clone())?;
-        }
-
-        let cont = Continuation::default();
-
-        let walker = GitWalker::new(DEFAULT_SOURCE_CVEPROJECT, ())
-            .continuation(cont)
-            .working_dir(path.clone())
-            .depth(3);
-
-        let _cont = walker.run().await.expect("should not fail");
-
-        let cont = git_reset(&path, "HEAD~2").expect("must not fail");
-
-        let walker = GitWalker::new(DEFAULT_SOURCE_CVEPROJECT, ())
-            .continuation(cont)
-            .working_dir(path);
-
-        walker.run().await.expect("should not fail");
-
-        Ok(())
-    }
-}
