+++ modules/importer/src/runner/osv/handler.rs
@@ -62,64 +62,3 @@ where
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod test {
-    use crate::runner::common::walker::{git_reset, Continuation, GitWalker};
-    use std::path::PathBuf;
-
-    #[test_log::test(tokio::test)]
-    async fn test_osv_walker() -> Result<(), anyhow::Error> {
-        const SOURCE: &str = "https://github.com/RConsortium/r-advisory-database";
-        let path = PathBuf::from(format!(
-            "{}target/test.data/test_walker.git",
-            env!("CARGO_WORKSPACE_ROOT")
-        ));
-        if path.exists() {
-            std::fs::remove_dir_all(path.clone())?;
-        }
-
-        let cont = Continuation::default();
-
-        let walker = GitWalker::new(SOURCE, ())
-            .path(Some("vulns"))
-            .continuation(cont)
-            .working_dir(path.clone())
-            .depth(3);
-
-        let _cont = walker.run().await.expect("should not fail");
-
-        let cont = git_reset(&path, "HEAD~2").expect("must not fail");
-
-        let walker = GitWalker::new(SOURCE, ())
-            .path(Some("vulns"))
-            .continuation(cont)
-            .working_dir(path);
-
-        walker.run().await.expect("should not fail");
-
-        Ok(())
-    }
-
-    /// ensure that using `path`, we can't escape the repo directory
-    #[test_log::test(tokio::test)]
-    async fn test_osv_walker_fail_escape() {
-        const SOURCE: &str = "https://github.com/RConsortium/r-advisory-database";
-        let path = PathBuf::from(format!(
-            "{}target/test.data/test_walker_fail_escape.git",
-            env!("CARGO_WORKSPACE_ROOT")
-        ));
-
-        let cont = Continuation::default();
-
-        let walker = GitWalker::new(SOURCE, ())
-            .path(Some("/etc"))
-            .continuation(cont)
-            .working_dir(path.clone());
-
-        let r = walker.run().await;
-
-        // must fail as we try to escape the repository root
-        assert!(r.is_err());
-    }
-}
